order date format modified, keys added to components, reorganized component files
