+++ client/components/userOrderHistory/Order.js
@@ -6,8 +6,10 @@ const Order = ({order}) => {
   return (
     <div>
       <h1> Order #{order.id}</h1>
-      <h2>Created on: {order.createdAt}</h2>
-      {order.orderProducts.map(el => <OrderProduct product={el.product} />)}
+      <h2>Ordered on: {new Date(order.createdAt).toString().slice(0, 15)}</h2>
+      {order.orderProducts.map((el, i) => (
+        <OrderProduct key={i} product={el.product} />
+      ))}
     </div>
   )
 }
