+++ client/components/userOrderHistory/index.js
@@ -1,0 +1,3 @@
+export {default} from './OrderHistory'
+export {default as Order} from './Order'
+export {default as OrderProduct} from './OrderProduct'
